added security

+++ src/main/java/com/innowise/userservice/service/impl/CardServiceImpl.java
@@ -5,9 +5,11 @@ import com.innowise.userservice.dao.UserRepository;
 import com.innowise.userservice.dto.CardDto;
 import com.innowise.userservice.dto.CreateCardDto;
 import com.innowise.userservice.dto.UpdateCardDto;
+import com.innowise.userservice.exception.AccessDeniedException;
 import com.innowise.userservice.exception.AccountIsNotActivatedException;
 import com.innowise.userservice.exception.BadIncomeDataException;
 import com.innowise.userservice.exception.EntityNotFoundException;
+import com.innowise.userservice.httpfilter.UserPrincipal;
 import com.innowise.userservice.mapper.CardMapper;
 import com.innowise.userservice.model.Card;
 import com.innowise.userservice.model.User;
@@ -33,8 +35,12 @@ public class CardServiceImpl implements CardService {
     @Override
     @Transactional
     @CacheEvict(value = "users", key = "#createCardDto.userId")
-    public CardDto createCard(CreateCardDto createCardDto) {
+    public CardDto createCard(CreateCardDto createCardDto, UserPrincipal principal) {
         User user = userRepository.findById(createCardDto.getUserId()).orElseThrow(() -> new EntityNotFoundException("User not found"));
+
+        if(!principal.getRole().equals("ADMIN") && !user.getId().equals(principal.getId())) {
+            throw new AccessDeniedException("Forbidden");
+        }
 
         if(!user.isActive()) {
             throw new AccountIsNotActivatedException("User account is not activated");
@@ -55,8 +61,13 @@ public class CardServiceImpl implements CardService {
     }
 
     @Override
-    public CardDto getCardById(Long id) {
+    public CardDto getCardById(Long id, UserPrincipal principal) {
         Card card = cardRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(CARD_NOT_FOUND_MESSAGE));
+
+        if(!principal.getRole().equals("ADMIN") && !card.getUser().getId().equals(principal.getId())) {
+            throw new AccessDeniedException("Forbidden");
+        }
+
         return cardMapper.toDto(card);
     }
 
@@ -89,8 +100,13 @@ public class CardServiceImpl implements CardService {
 
     @Override
     @Transactional
-    public void deleteCard(Long id) {
+    public void deleteCard(Long id, UserPrincipal principal) {
         Card card = cardRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(CARD_NOT_FOUND_MESSAGE));
+
+        if(!principal.getRole().equals("ADMIN") && !card.getUser().getId().equals(principal.getId())) {
+            throw new AccessDeniedException("Forbidden");
+        }
+
         cardRepository.delete(card);
         cacheManager.getCache("users")
                 .evict(card.getUser().getId());
